+++ tests/test_count.py
@@ -1,9 +1,12 @@
-from recurtools import countrecursive
+from recurtools import nested
 
 # https://stackoverflow.com/questions/72321903/deep-list-count-count-lists-within-lists
 
 def test_count_ints():
-    assert countrecursive([1, 2, [3, 2]],2) == 2
+    assert nested([1, 2, [3, 2]]).count(2) == 2
 
 def test_count_strs():
-    assert countrecursive(["ab", "b", ["c", "db", ["e","bob"]]],"b") == 5
+    assert nested(["ab", "b", ["c", "db", ["e","bob"]]]).count("b") == 5
+
+def test_count_notfound():
+    assert nested([1, 2, [3, 2]]).count(4) == 0